Improve find account with balance function.
Before we listed all the accounts in the database and then
we just picked the once with balance. This is very bad for
the performance if the database is huge. So instead we pick
a random account until we get one with a balance. This is much
faster for most of the times.