Refactor sign_tx_client.
Remove the cb since we want to be able to add new callback for each
transaction. Before we could only add one callback for all transactions.
We have made the Consumer class as daemon so that when the main thread
is done we want the Consumer thread to die too.