Add a dummy wallet and remove global wallet_counter.
Since we can only a add wallet asyncronous we must add a placeholder
wallet while waiting for the public key from the airgapped computer.
Once we get the public key we replace the dummy wallet with the real.
one. If we don't have a dummy wallet we will not have the correct wallet
count while waiting for the public key.