Feature: Use a timout to find minimodem blocks.
Sometimes minimodem hangs for some unknown reason which leads to
the whole signing process gets blocked.
To solve this issue we have added a timeout in the select function.
If we do get a timeout we will kill the receiver thread and restart
it. We will also resend the transaction.