Refactor: Move most of the calls insight away from wallet.
Wallet object used to call insight each time it was created. Once the wallet
got many public addresses this had huge impact on performance. To solve this
problem we moved most of the calls to insight from wallet to blockchainfetcher.
The wallet instead saved it's state in a json format so when a new wallet object
was called we used all the key information previously saved in the database.
Instead of calling insight for all public keys.